# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Band Manager API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /merch-booth/checkouts/cash:
    post:
      operationId: createCashCheckout
      summary: Finalize an online cash checkout
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashCheckoutRequest'
      responses:
        '201':
          description: Cash checkout finalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          description: Invalid checkout request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Role cannot checkout in alpha
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booth item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Insufficient stock or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /merch-booth/checkouts/pix:
    post:
      operationId: createPixCheckout
      summary: Reserve inventory and create a Pix payment QR code
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PixCheckoutRequest'
      responses:
        '201':
          description: Pix checkout created and awaiting payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          description: Invalid checkout request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Role cannot checkout in alpha
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booth item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Insufficient stock or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Payment provider request failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /merch-booth/checkouts/card:
    post:
      operationId: createCardCheckout
      summary: Reserve inventory and create a MercadoPago Point card payment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardCheckoutRequest'
      responses:
        '201':
          description: Card checkout created and awaiting terminal payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          description: Invalid checkout request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Role cannot checkout in alpha
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booth item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Insufficient stock or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Payment provider request failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /webhooks/mercadopago/orders:
    post:
      operationId: handleMercadoPagoOrderWebhook
      summary: Receive signed MercadoPago order webhook notifications
      security: []
      parameters:
        - name: data.id
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum:
              - order
        - name: x-signature
          in: header
          required: true
          schema:
            type: string
        - name: x-request-id
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '401':
          description: Invalid webhook signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Payment provider request failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
      description: Unique key scoped by band and operation for mutating requests.
  schemas:
...
    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: integer
          minimum: 0
        currency:
          type: string
          enum:
            - BRL
    SignedMoney:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: integer
        currency:
          type: string
          enum:
            - BRL
...
    InventoryCategory:
      type: string
      enum:
        - shirt
        - hoodie
        - tote_bag
        - patch
        - sticker
        - vinyl
        - cd
        - cassette
        - accessory
    InventorySize:
      type: string
      enum:
        - not_applicable
        - one_size
        - pp
        - p
        - m
        - g
        - gg
        - xgg
...
    CashCheckoutRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CashCheckoutItem'
    PixCheckoutRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CashCheckoutItem'
    CardCheckoutRequest:
      type: object
      required:
        - items
        - cardType
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CashCheckoutItem'
        cardType:
          type: string
          enum:
            - credit_card
            - debit_card
...
    WebhookResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - processed
    CashCheckoutItem:
      type: object
      required:
        - variantId
        - quantity
      properties:
        variantId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
    Sale:
      type: object
      required:
        - id
        - bandId
        - status
        - total
        - expectedProfit
        - items
        - payment
        - transactions
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        bandId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - pending_payment
            - finalized
            - canceled
        total:
          $ref: '#/components/schemas/Money'
        expectedProfit:
          $ref: '#/components/schemas/SignedMoney'
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleItem'
        payment:
          $ref: '#/components/schemas/Payment'
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SaleItem:
      type: object
      required:
        - id
        - saleId
        - productId
        - variantId
        - productName
        - category
        - size
        - colour
        - quantity
        - unitPrice
        - unitCost
        - lineTotal
        - expectedProfit
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        saleId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        variantId:
          type: string
          format: uuid
        productName:
          type: string
        category:
          $ref: '#/components/schemas/InventoryCategory'
        size:
          $ref: '#/components/schemas/InventorySize'
        colour:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          $ref: '#/components/schemas/Money'
        unitCost:
          $ref: '#/components/schemas/Money'
        lineTotal:
          $ref: '#/components/schemas/Money'
        expectedProfit:
          $ref: '#/components/schemas/SignedMoney'
        createdAt:
          type: string
          format: date-time
    Payment:
      type: object
      required:
        - id
        - saleId
        - method
        - status
        - amount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        saleId:
          type: string
          format: uuid
        method:
          type: string
          enum:
            - cash
            - pix
            - card
        status:
          type: string
          enum:
            - provider_pending
            - action_required
            - processing
            - confirmed
            - failed
            - canceled
            - expired
        amount:
          $ref: '#/components/schemas/Money'
        provider:
          type: string
        providerOrderId:
          type: string
        providerPaymentId:
          type: string
        providerReferenceId:
          type: string
        externalReference:
          type: string
        providerStatus:
          type: string
        providerStatusDetail:
          type: string
        expiresAt:
          type: string
          format: date-time
        pixQrCode:
          type: string
        pixQrCodeBase64:
          type: string
        pixTicketUrl:
          type: string
        pointTerminalId:
          type: string
        cardPaymentType:
          type: string
          enum:
            - credit_card
            - debit_card
        cardInstallments:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Transaction:
      type: object
      required:
        - id
        - saleId
        - saleItemId
        - amount
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        saleId:
          type: string
          format: uuid
        saleItemId:
          type: string
          format: uuid
        amount:
          $ref: '#/components/schemas/Money'
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string